\docType{function}
\name{pclookup.release}
\alias{pclookup.release}
\title{GB Postcode Lookup Table Version}
\format{A character variable stating year and month of release}
\source{
  \url{http://www.ons.gov.uk/ons/guide-method/geography/products/postcode-directories/-nspp-/index.html}
}
\usage{
  pclookup.release()
}
\description{
  A function stating the year and month of release of this
  version of the GB postcode lookup table
}
\examples{
pclookup.release()
}

